+++ CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [3.0.2] - Ongoing
+This update introduces minor bug fixes and improvements
 
-This is an ongoing update
+### Changed
+ - ncc now supports installing packages using `ncc --install=<package-name>` which will install the package the same way
+   as `ncc install --package=<package-name>` does.
+
+### Fixed
+ - Fixed issue where `--reinstall` is not respected when installing local packages using `ncc install --local=<path-to-package>`
 
 
 ## [3.0.1] - 2026-01-05